removed another couple of mallocs and longs

+++ src/SoundBuffer.cpp
@@ -288,7 +288,7 @@ bool SoundBuffer::_loadOgg(Stream* source) {
 		}
 
 		ogg_int64_t byteSize = (pcmEnd - pcmStart) * wordSize * info->channels;
-		mChunks.emplace_back(make_unique<Chunk>(this, (long)fileStart, (long)byteSize));
+		mChunks.emplace_back(make_unique<Chunk>(this, fileStart, byteSize));
 
 		pcmStart = pcmEnd;
 		fileStart = fileEnd;
